Fix waiting for kubelet restart
As kubelet defaults to CRI in 1.6, it no longer kills pods after
restarts, so just sleeping is not an option and we must poll /configz to
make sure kubelet has restarted. In case if CRI is switched off, the
bootstrap procedure should still work as it's still compatible with pod
restart.