Demo.sh script now uses ssh instead virsh console
Since we have disabled virsh console in favor of VM logging,
demo.sh script was unable to connect to the cirros VM at the
very end of the deployment process. Therefore we update script
so that it connects to the VM with ssh instead of virsh console.

There are two workarounds used in this patch:

1. password authentication is used
2. infinite while loop is used

The former is required since we need demo.sh script to be self-contained
so we are unable to ship it with ssh keys. The latter is required since
ssh server in cirros is started with some delay after cirros is already
running.

The easiest way to wait for ssh server to start was to simply retry in
while loop until it blocks. This brings some minor issue that once user
cancels ssh connection with CTRL + D, script attempts to ssh again. But
IMO this is not an issue, since user can simply hit CTRL + C if she doesn't
want to connect again and the script will finally exit.

Fixes https://github.com/Mirantis/virtlet/issues/304

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>