Demo.sh script now uses ssh instead virsh console
Since we have disabled virsh console in favor of VM logging,
demo.sh script was unable to connect to the cirros VM at the
very end of the deployment process. Therefore we update script
so that it connects to the VM with ssh instead of virsh console.

Fixes https://github.com/Mirantis/virtlet/issues/304

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>