Follow upstream changes of libvirt-xml-go
Virtlet uses a patch to adjust libvirt-xml-go domain definition.
But due to changes in libvirt-xml-go upstream it's no longer possible
to apply the patch. With this commit we update the patch so that it
now follows the upstream. Also, some changes were conflicting therefore
we needed to modify the Virtlet code that relies on those changes.

Namely, virtlet's patch suggested following struct for providing bootcmd:
```
DomainCMDLine
|-> Args: []QemuArg
|-> Envs: []QemuEnv
```

but the upstream code now provides following struct:
```
DomainQEMUCommandline
|-> Args: []DomainQEMUCommandlineArg
|-> Envs: []DomainQEMUCommandlineEnv
```

So we had to enforce such naming also in Virtlet code.

Fixes https://github.com/Mirantis/virtlet/issues/375
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>