Fix cni conflist handling
This is done by following current kubelet behavior more closely.
The problem became apparent with Flannel 0.9.1 update which
made use of 'portmap' plugin which is configured in a single
conflist file with the main 'flannel' plugin.
As portmap is needed for hostPort feature, there's high possibility that
other CNIs could become broken soon.