Provide symlinks & fix mounts for block volumes
The user can use `mounts` in cloud-init userdata
to mount the volumes, but there's one problem.
The symlink script that ensures that the "container"
device paths lead to the actual block devices corresponding
to the PVs runs after cloud-init mounts are handled.
Thus, we need to update cloud-init mounts, replacing
the references to PV block devices with `/dev/disk/by-path/...`