Update k8s example
The example is updated to make it use "modern" StatefulSet definition
and for compatiblity with CoreDNS in k8s 1.12. Previously, the example
did work except for the very final stage where kubeconfig is set up
for the root user, which was because the script was waiting for DNS
pods to become "ready 3/3" instead of "ready 1/1" which must be used
for CoreDNS. Also, some unneeded options are removed, and ephemeral
flexvolume usage is eliminated in favor of `VirtletRootVolumeSize`.