Make Virtlet work with newer Calico
For now, only DHCP-based networking works for Calico.
So, if your VM supports cloud-init, you must either set
`VirtletForceDHCPNetworkConfig: "true"` annotation for it
or add a script that will add a link-scoped route
to `169.254.1.1` for the CNI-provided link, as well
as the default route via `169.254.1.1`.

The patch is inspired by kubevirt/kubevirt#1436 and removes
the old scheme that was making a "fake pod", because it's no
longer possible to create such "fake pods" in new versions
of Calico (they always must correspond to a real pod).

The remaining difficulty is that it's apparently impossible
to specify the link-scoped routes through cloud-init network
config.